updated e2e test
